Multithreaded the core loop
