+++ Engine/Core/Game/GameBase.cpp
@@ -1,36 +1,33 @@
 #include "GameBase.h"
-#include "Common/Defines.h"
-#include "Core/Application/Window.h"
 #include <format>
 
 namespace Clone::Game
 {
-	GameBase::GameBase():
-		m_parentWindow(nullptr),
-		m_engine()
+	GameBase::GameBase()
+		:
+		m_appConfig()
 	{
 	}
 
-	bool GameBase::PreInit(Application::Window* parentWindow)
+	bool GameBase::PreInit(Windowing::WindowPtr parentWindow)
 	{
 		CLONE_DEBUG(GameBase, "Started pre-Initializing Game Base");
 
-		m_parentWindow = parentWindow;
-		//m_engine.Engine = std::make_unique<Engine::Engine>(m_parentWindow->GetHandle());
 				
 		CLONE_DEBUG(GameBase, "Finishd pre-Initializing Game Base");
 		return true;
 	}
 
-	void GameBase::PreUpdate(double deltaTime)
+	void GameBase::PreUpdate(double deltaTime, const Input::Event& e)
 	{
 	}
 
 	void GameBase::Render()
 	{
-		GetRenderer()->IsVsyncEnabled = false;
-		GetRenderer()->Clear(0.1f, 0.1f, 0.15f, 1.0f);
-		GetRenderer()->RenderFrame();
+		//GetRenderer()->IsVsyncEnabled = false;
+		//GetRenderer()->Clear(0.1f, 0.1f, 0.15f, 1.0f);
+		//GetRenderer()->RenderFrame();
+		CLONE_WARN(Game Base, "Rendering!");
 	}
 
 	void GameBase::PreShutdown()
