+++ Engine/Core/Windowing/Window.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "WindowDesc.h"
-#include "EventQueue.h"
+#include "Core/Input/EventQueue.h"
 #include <memory>
 #include <functional>
 #include <unordered_map>
@@ -11,7 +11,7 @@ namespace Clone::Windowing
 {
 	class Window
 	{
-		friend class EventQueue;
+		friend class Input::InputEventQueue;
 	public:
 		enum class HitRectType : size_t
 		{
@@ -36,7 +36,7 @@ namespace Clone::Windowing
 		Window();
 		~Window();
 
-		bool Create(HINSTANCE hInst, const WindowDesc& desc, EventQueue& eventQueue);
+		bool Create(HINSTANCE hInst, const WindowDesc& desc, Input::InputEventQueue& eventQueue);
 
 		const std::string GetTitle() const;
 		void SetTitle(const std::string& title);
@@ -55,10 +55,11 @@ namespace Clone::Windowing
 		unsigned GetBackgroundColor() const;
 		void SetBackgroundColor(unsigned color) ;
 
+		void Show() const;
 		void Minimize() const;
 		void Maximize() const;
 		void Close();
-		void TrackEventAsync(const std::function<void(const Event)>& func);
+		void TrackEventAsync(const std::function<void(const Input::Event)>& func);
 
 		void SetTaskbarProgress(float progress);
 		
@@ -71,28 +72,28 @@ namespace Clone::Windowing
 		std::vector<HitRect> HitRects;
 
 	protected:
-		void ExecuteEventCallback(const Event e);
-		std::function<void(const Event)> m_callback;
+		void ExecuteEventCallback(const Input::Event e);
+		std::function<void(const Input::Event)> m_callback;
 		static LRESULT CALLBACK WindowProcStatic(HWND hwnd, UINT msg, WPARAM wparam, LPARAM lparam);
 
 		LRESULT WindowProc(UINT msg, WPARAM wparam, LPARAM lparam);
 
-		HWND           m_hWnd{ nullptr };
-		HINSTANCE      m_hInstance{ nullptr };
-		EventQueue*    m_eventQueue{ nullptr };
-		WindowDesc     m_desc;
-		WNDCLASSEX     m_wndClass{ 0 };
-		RECT           m_windowRect{ 0 };
-		DEVMODE        m_screenSettings{ 0 };
-		DWORD          m_exStyle{ 0x00 };
-		DWORD          m_style{ 0x00 };
-		ITaskbarList3* m_taskbarList{ nullptr };
-		unsigned       m_backgroundColor{ 0xFFFFFFFF };
+		HWND                    m_hWnd{ nullptr };
+		HINSTANCE               m_hInstance{ nullptr };
+		Input::InputEventQueue* m_eventQueue{ nullptr };
+		WindowDesc              m_desc{};
+		WNDCLASSEX              m_wndClass{ 0 };
+		RECT                    m_windowRect{ 0 };
+		DEVMODE                 m_screenSettings{ 0 };
+		DWORD                   m_exStyle{ 0x00 };
+		DWORD                   m_style{ 0x00 };
+		ITaskbarList3*          m_taskbarList{ nullptr };
+		unsigned                m_backgroundColor{ 0xFFFFFFFF };
 	};
 
 	static thread_local Window* _windowBeingCreated = nullptr;
 	static thread_local std::unordered_map<HWND, Window*> _hWndMap = {};
 
-	using WindowPtr = std::shared_ptr<Window>;
-	using WindowWeakPtr = std::weak_ptr<Window>;
+	using WindowPtr = std::shared_ptr<Clone::Windowing::Window>;
+	using WindowWeakPtr = std::weak_ptr<Clone::Windowing::Window>;
 }
