+++ Runtime/Main.cpp
@@ -12,6 +12,7 @@ int WINAPI wWinMain(
 	_In_ int nShowCmd)
 {
 	Application app(hInstance);
+	app.Init();
 	app.Run();
 
 	return 0;
