+++ Testbench/src/Testbench.cpp
@@ -2,21 +2,28 @@
 
 TestbenchClass::TestbenchClass() : Game::GameBase()
 {
+	// Configure application here
+	auto& config = GetAppConfig();
+	config.WindowDesc.Title = "Testbench ";
+	config.WindowDesc.CanFullscreen = false;
+	config.WindowDesc.IsResizable = true;
+	config.WindowDesc.Width = 1280;
+	config.WindowDesc.Height = 720;
 }
 
 bool TestbenchClass::Init()
 {
 	CLONE_DEBUG(Testbench, "Started Testbench initialization");
-	
-	
+		
 	CLONE_DEBUG(Testbench, "Finished Testbench initialization");
 	return true;
 }
 
-void TestbenchClass::Update(double deltaTime)
+void TestbenchClass::Update(double deltaTime, const Input::Event& e)
 {
+	CLONE_INFO(Testbench, "Updating!");
 }
-
+ 
 void TestbenchClass::Shutdown()
 {
 	CLONE_DEBUG(Testbench, "Testbench Shutdown");
